+++ src/components/AppHeader/AppHeader.tsx
@@ -12,7 +12,7 @@ export const AppHeader = () => {
     <header className={appHeader.header}>
       <nav className="pt-4 pb-4">
         <ul>
-          <li className="pt-4 pb-4 pr-5 pl-5 mr-2">
+          <li className="pt-4 pb-4 pr-5 mr-2">
             <NavItem name="Конструктор" link="/">
               <BurgerIcon type="primary" />
             </NavItem>
