--- src/pages/ProfileOrders/profile-orders.module.css
@@ -1,4 +1,0 @@
-.container {
-  display: flex;
-  width: 480px;
-}
