--- src/pages/ProfileOrders/profile-orders.tsx
@@ -1,5 +1,0 @@
-import style from './profile-orders.module.css'
-
-export const ProfileOrders = () => {
-  return <div className={style.container}>PROFILE-ORDERS</div>
-}
