+++ src/pages/ProfilePage/profile-page.module.css
@@ -1,8 +1,10 @@
 .wrapper {
   display: flex;
   height: 100%;
-  justify-content: center;
-  margin-top: 120px;
+  width: 1240px;
+  justify-content: flex-start;
+  margin: 120px auto 0;
+  box-sizing: border-box;
 }
 
 .left-side,
